Share configuration between environments.
