+++ Content/default/webpack.config.js
@@ -1,19 +1,8 @@
-// Template for webpack.config.js in Fable projects
-// Find latest version in https://github.com/fable-compiler/webpack-config-template
-
-// In most cases, you'll only need to edit the CONFIG object (after dependencies)
-// See below if you need better fine-tuning of Webpack options
-
-const path = require('path');
-const HtmlWebpackPlugin = require('html-webpack-plugin');
-const CopyWebpackPlugin = require('copy-webpack-plugin');
-const MiniCssExtractPlugin = require('mini-css-extract-plugin')
-
-// The NODE_ENV is passed to npm as an argument from the command line
-// for example: npm run start --node-env=development
-process.env.NODE_ENV = process.env.NODE_ENV ?? "development";
-const mode = process.env.NODE_ENV;
-const isProduction = mode === 'production';
+// In most cases, you'll only need to edit the CONFIG/TEST_CONFIG objects
+// CONFIG is the configuration used to run the application.
+// TEST_CONFIG is the configuration used to run tests.
+// If you need better fine-tuning of Webpack options check the buildConfig function.
+const common = require("./webpack.common");
 
 const CONFIG = {
     // The tags to include the generated JS and CSS will be automatically injected in the HTML template
@@ -39,102 +28,26 @@ const CONFIG = {
     }
 }
 
-console.log('Bundling for ' + mode + '...');
+const TEST_CONFIG = {
+    // The tags to include the generated JS and CSS will be automatically injected in the HTML template
+    // See https://github.com/jantimon/html-webpack-plugin
+    indexHtmlTemplate: 'tests/Client/index.html',
+    fsharpEntry: 'tests/Client/output/Client.Tests.js',
+    outputDir: 'tests/Client',
+    assetsDir: 'tests/Client',
+    devServerPort: 8081,
+    // When using webpack-dev-server, you may need to redirect some calls
+    // to a external API server. See https://webpack.js.org/configuration/dev-server/#devserver-proxy
+    devServerProxy: undefined,
+}
 
-// The HtmlWebpackPlugin allows us to use a template for the index.html page
-// and automatically injects <script> or <link> tags for generated bundles.
-const commonPlugins = [
-    new HtmlWebpackPlugin({
-        filename: 'index.html',
-        template: resolve(CONFIG.indexHtmlTemplate)
-    })
-];
+module.exports = function(env, arg) {
+    // environment variables docs: https://webpack.js.org/api/cli/#environment-options
+    const config = env.test ? TEST_CONFIG : CONFIG;
+    // Flags docs: https://webpack.js.org/api/cli/#flags
+    const mode = arg.mode ?? 'development';
 
-module.exports = {
-    // In development, split the JavaScript and CSS files in order to
-    // have a faster HMR support. In production bundle styles together
-    // with the code because the MiniCssExtractPlugin will extract the
-    // CSS in a separate files.
-    entry: {
-        app: resolve(CONFIG.fsharpEntry)
-    },
-    // Add a hash to the output file name in production
-    // to prevent browser caching if code changes
-    output: {
-        path: resolve(CONFIG.outputDir),
-        publicPath: '/',
-        filename: isProduction ? '[name].[contenthash].js' : '[name].js'
-    },
-    mode: mode,
-    devtool: isProduction ? 'source-map' : 'eval-source-map',
-    optimization: {
-        runtimeChunk: 'single',
-        moduleIds: 'deterministic',
-        // Split the code coming from npm packages into a different file.
-        // 3rd party dependencies change less often, let the browser cache them.
-        splitChunks: {
-            chunks: 'all'
-        }
-    },
-    // Besides the HtmlPlugin, we use the following plugins:
-    // PRODUCTION
-    //      - MiniCssExtractPlugin: Extracts CSS from bundle to a different file
-    //          To minify CSS, see https://github.com/webpack-contrib/mini-css-extract-plugin#minimizing-for-production
-    //      - CopyWebpackPlugin: Copies static assets to output directory
-    // DEVELOPMENT
-    //      - HotModuleReplacementPlugin: Enables hot reloading when code changes without refreshing
-    plugins: isProduction ?
-        commonPlugins.concat([
-            new MiniCssExtractPlugin({ filename: 'style.[name].[contenthash].css' }),
-            new CopyWebpackPlugin({ patterns: [{ from: resolve(CONFIG.assetsDir) }] }),
-        ])
-        : commonPlugins,
-    resolve: {
-        // See https://github.com/fable-compiler/Fable/issues/1490
-        symlinks: false
-    },
-    // Configuration for webpack-dev-server
-    devServer: {
-        static: {
-            directory: resolve(CONFIG.assetsDir),
-            publicPath: '/'
-        },
-        host: '0.0.0.0',
-        port: CONFIG.devServerPort,
-        proxy: CONFIG.devServerProxy,
-        hot: true,
-        historyApiFallback: true
-    },
-    // - sass-loaders: transforms SASS/SCSS into JS
-    // - file-loader: Moves files referenced in the code (fonts, images) into output folder
-    module: {
-        rules: [
-            {
-                test: /\.(sass|scss|css)$/,
-                use: [
-                    isProduction
-                        ? MiniCssExtractPlugin.loader
-                        : 'style-loader',
-                    'css-loader',
-                    {
-                        loader: 'sass-loader',
-                        options: { implementation: require('sass') }
-                    }
-                ],
-            },
-            {
-                test: /\.(png|jpg|jpeg|gif|svg|woff|woff2|ttf|eot)(\?.*)?$/,
-                use: ['file-loader']
-            },
-            {
-                test: /\.js$/,
-                enforce: 'pre',
-                use: ['source-map-loader'],
-            }
-        ]
-    }
-};
+    console.log(`Bundling for ${env.test ? 'test' : 'run'} - ${mode} ...`);
 
-function resolve(filePath) {
-    return path.isAbsolute(filePath) ? filePath : path.join(__dirname, filePath);
+    return common.buildConfig(config, mode);
 }
